requirements.txt
langsmith
langgraph
langchain
langchain-ollama
pydantic
requests
python-dotenv
amadeus
typing_extensions
pytest
ruff
black
